+++ src/utilities/hash_map.h
@@ -40,9 +40,9 @@ static inline size_t hash_djb2(const char *string) {
 }
 
 static inline size_t hash_djb2_unterminated(const char *string, size_t len) {
-    fprintf(stderr, "hash_djb2_unterminated got: '%.*s', length: %zu.", len, string, len);
+    fprintf(stderr, "hash_djb2_unterminated got: '%.*s', length: %zu.", (int)len, string, len);
     size_t hash = 5381;
-    for (int i = 0; i < len; i++)
+    for (size_t i = 0; i < len; i++)
         hash = ((hash << 5) + hash) + (unsigned char)string[i];
     fprintf(stderr, " Hashed as: %zu\n", hash);
     return hash;
